Move group check to its own package
